Backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pymongo>=4.6.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
google-genai>=0.2.0
groq>=0.4.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
